split server and timings into two files, server now calls timings
replace webserver with url in console. no listening required

use server to call fetch_timings, then delete file generated by fetch timings and kill process
