+++ setup.sh
@@ -5,7 +5,7 @@ sudo mkdir app
 cd /app/
 
 # GET PHANTOMJS
-sudo apt-get install build-essential chrpath libssl-dev libxft-dev libfreetype6-dev libfreetype6 libfontconfig1-dev libfontconfig1 -y
+sudo apt-get install build-essential chrpath libssl-dev libxft-dev libfreetype6-dev libfreetype6 libfontconfig1-dev libfontconfig1 uuid -y
 sudo wget https://bitbucket.org/ariya/phantomjs/downloads/phantomjs-2.1.1-linux-x86_64.tar.bz2 -O /app/phantomjs.tar.bz2
 sudo tar xvjf /app/phantomjs.tar.bz2 -C /app/
 
@@ -18,5 +18,5 @@ curl -sL https://deb.nodesource.com/setup_11.x | sudo -E bash -
 sudo apt-get install -y nodejs
 
 # KICK APP IN BACKGROUND
-sudo /app/phantomjs-2.1.1-linux-x86_64/bin/phantomjs /app/server.js &
+sudo node /app/server.js &
 EOF
